Create repeatstring.js
This is a function repeatStringNumTimes() that take in two parameters. (1.) the str (string) that is to be repeated (2.) the num (number), number of times that it will be repeat. And the str was returned after calling the function with the two arguments just to be sure that we can access our arguments in the function